removed wrapper component. moved styles to global level
