--- src/app/public/modules/angular-tree/angular-tree-root.component.html
@@ -1,14 +1,0 @@
-<div class="sky-angular-tree-root">
-
-  <sky-angular-tree-toolbar *ngIf="showToolbar"
-    [showSelectButtons]="treeComponent.treeModel.options.useCheckbox"
-    (clearAllClick)="onClearAllClick()"
-    (collapseAllClick)="onCollapseAllClick()"
-    (expandAllClick)="onExpandAllClick()"
-    (selectAllClick)="onSelectAllClick()"
-  >
-  </sky-angular-tree-toolbar>
-
-  <ng-content select="tree-root"></ng-content>
-
-</div>
